numpy shenanigains

+++ verl/trainer/ppo/metric_utils.py
@@ -233,6 +233,7 @@ def compute_data_metrics(
 
     # log reward_extra_infos_dict means for training
     for key, values in reward_extra_infos_dict.items():
+        values = values.astype("float64")
         metrics[f"critic/rewards/{key}/min"] = np.nanmin(values)
         metrics[f"critic/rewards/{key}/max"] = np.nanmax(values)
         metrics[f"critic/rewards/{key}/mean"] = np.nanmean(values)
@@ -246,7 +247,11 @@ def compute_data_metrics(
             metrics[f"redel/{redel_key}/mean"] = val.mean()
     if "n_zombie" in batch.non_tensor_batch and "n_children" in batch.non_tensor_batch:
         n_children = batch.non_tensor_batch["n_children"]
-        zombie_ratio = np.where(n_children == 0, np.nan, batch.non_tensor_batch["n_zombie"] / np.where(n_children == 0, 1, n_children))
+        zombie_ratio = np.where(
+            n_children == 0,
+            np.nan,
+            batch.non_tensor_batch["n_zombie"] / np.where(n_children == 0, 1.0, n_children),
+        ).astype("float64")
         metrics["redel/zombie_ratio/nanmin"] = np.nanmin(zombie_ratio)
         metrics["redel/zombie_ratio/nanmax"] = np.nanmax(zombie_ratio)
         metrics["redel/zombie_ratio/nanmean"] = np.nanmean(zombie_ratio)
